+++ src/pudl/metadata/classes.py
@@ -1296,6 +1296,7 @@ class Resource(PudlMeta):
             "pudl",
             "nrelatb",
             "vcerare",
+            "phmsagas",
             "sec",
         ]
         | None
@@ -1325,6 +1326,7 @@ class Resource(PudlMeta):
             "service_territories",
             "nrelatb",
             "vcerare",
+            "phmsagas",
             "sec10k",
         ]
         | None