+++ src/pudl/metadata/fields.py
@@ -57,6 +57,10 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "description": "Demand reduction actually achieved by demand response activities. Measured at the time of the company's annual system peak hour.",
         "unit": "MW",
     },
+    "additional_information": {
+        "type": "string",
+        "description": "Any additional information which will assist in clarifying or classifying the reported data.",
+    },
     "additions": {
         "type": "number",
         "description": "Cost of acquisition of items classified within the account.",
@@ -1559,6 +1563,54 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "description": "Whether the reported technology data is estimated or actual.",
         "constraints": {"enum": list(ESTIMATED_OR_ACTUAL.values())},
     },
+    "excavation_damage_excavation_practices": {
+        "type": "integer",
+        "description": (
+            "Number of incidents in which damages resulted from failure to maintain "
+            "marks; or failure to support exposed facilities; or failure to use hand "
+            "tools where required; or failure to test-hole (pot-hole); or improper "
+            "backfilling practices; or failure to maintain clearance; or other "
+            "insufficient excavation practices."
+        ),
+    },
+    "excavation_damage_locating_practices": {
+        "type": "integer",
+        "description": (
+            "Number of incidents in which damages resulted from facility could not be "
+            "found or located; or facility marking or location not sufficient; or "
+            "facility was not located or marked; or incorrect facility records/maps."
+        ),
+    },
+    "excavation_damage_one_call_notification": {
+        "type": "integer",
+        "description": (
+            "Number of incidents in which damages resulted from no notification made "
+            "to the One-Call Center; or notification to one-call center made, but not "
+            "sufficient; or wrong information provided to One Call Center"
+        ),
+    },
+    "excavation_damage_other": {
+        "type": "integer",
+        "description": (
+            "Number of incidents in which damages resulted from from One-Call Center "
+            "error; or abandoned facility; or deteriorated facility; or previous "
+            "damage or data not collected; or other."
+        ),
+    },
+    "excavation_damage_total": {
+        "type": "integer",
+        "description": (
+            "Total number of incidents resulting in damage resulting from any cause."
+        ),
+    },
+    "excavation_tickets": {
+        "type": "integer",
+        "description": (
+            "Number of Excavation Tickets received by the operator during the year, "
+            "(i.e., receipt of information by the operator from the notification "
+            "center)."
+        ),
+    },
     "exchange_energy_delivered_mwh": {
         "type": "number",
         "description": (
@@ -1583,6 +1635,13 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         },
     },
     "expense_type": {"type": "string", "description": "The type of expense."},
+    "federal_land_leaks_repaired_or_scheduled": {
+        "type": "integer",
+        "description": (
+            "Total number of leaks repaired, eliminated, or scheduled for repair on "
+            "federal land during the reporting year."
+        ),
+    },
     "ferc1_generator_agg_id": {
         "type": "integer",
         "description": "ID dynamically assigned by PUDL to EIA records with multiple matches to a single FERC ID in the FERC-EIA manual matching process.",
@@ -2181,6 +2240,29 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "number",
         "description": "Average power in megawatts delivered during time interval measured.",
         "unit": "MW",
+    },
+    "headquarters_city": {
+        "type": "string",
+        "description": "City where an operator's headquarters are located.",
+    },
+    "headquarters_county": {
+        "type": "string",
+        "description": "County where an operator's headquarters are located.",
+    },
+    "headquarters_state": {
+        "type": "string",
+        "description": "State where an operator's headquarters are located.",
+    },
+    "headquarters_street_address": {
+        "type": "string",
+        "description": "Street address for an operator's headquarters.",
+    },
+    "headquarters_zip": {
+        "type": "string",
+        "description": "Zipcode where an operator's headquarters are located.",
+        "constraints": {
+            "pattern": r"^\d{5}$",
+        },
     },
     "heat_content_mmbtu": {
         "type": "number",
@@ -3205,6 +3287,29 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "integer",
         "description": "Number of circuits in a transmission line.",
     },
+    "office_city": {
+        "type": "string",
+        "description": "City where an operator's office is located.",
+    },
+    "office_county": {
+        "type": "string",
+        "description": "County where an operator's office is located.",
+    },
+    "office_state": {
+        "type": "string",
+        "description": "State where an operator's office is located.",
+    },
+    "office_street_address": {
+        "type": "string",
+        "description": "Street address of an operator's office.",
+    },
+    "office_zip": {
+        "type": "string",
+        "description": "Zipcode where an operator's office is located.",
+        "constraints": {
+            "pattern": r"^\d{5}$",
+        },
+    },
     "oil_fraction_cost": {
         "type": "number",
         "description": "Oil cost as a percentage of overall fuel cost.",
@@ -3259,6 +3364,14 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
     "operator_utility_id_eia": {
         "type": "integer",
         "description": "The EIA utility Identification number for the operator utility.",
+    },
+    "operator_id_phmsa": {
+        "type": "integer",
+        "description": "PHMSA unique operator ID.",
+    },
+    "operator_name_phmsa": {
+        "type": "string",
+        "description": "PHMSA operator name.",
     },
     "opex_allowances": {"type": "number", "description": "Allowances.", "unit": "USD"},
     "opex_boiler": {
@@ -3572,6 +3685,17 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         # TODO Disambiguate column names. Usually this is over 60 minutes, but in
         # other tables it's not specified.
     },
+    "unaccounted_for_gas_fraction": {
+        "type": "number",
+        "description": (
+            "Unaccounted for gas as a fraction of total consumption "
+            "for the 12 months ending June 30 of the reporting year. "
+            "Calculated as follows: "
+            "Take the sum of: purchased gas + produced gas minus customer use + company use "
+            "+ appropriate adjustments. Then divide by the sum of customer use + company use "
+            "+ appropriate adjustments."
+        ),
+    },
     "percent_dry_cooling": {
         "description": "Percent of cooling load served by dry cooling components",
         "type": "number",
@@ -3783,6 +3907,26 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "number",
         "unit": "MW",
     },
+    "preparer_email": {
+        "type": "string",
+        "description": "Email address of representative who filed report.",
+    },
+    "preparer_fax": {
+        "type": "string",
+        "description": "Fax number of representative who filed report.",
+    },
+    "preparer_name": {
+        "type": "string",
+        "description": "Name of representative who filed report.",
+    },
+    "preparer_phone": {
+        "type": "string",
+        "description": "Phone number of representative who filed report.",
+    },
+    "preparer_title": {
+        "type": "string",
+        "description": "Title of representative who filed report.",
+    },
     "previously_canceled": {
         "type": "boolean",
         "description": "Indicates whether the generator was previously reported as indefinitely postponed or canceled",
@@ -3961,6 +4105,15 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "string",
         "description": "Timezone used by the reporting entity. For use in localizing UTC times.",
         "constraints": {"enum": US_TIMEZONES},
+    },
+    "report_number": {
+        "type": "integer",
+        "description": "Report number.",
+    },
+    "report_submission_type": {
+        "type": "string",
+        "description": "Type of report submitted, either Initial or Supplemental.",
+        "constraints": {"enum": ["Initial", "Supplemental"]},
     },
     "report_year": {
         "type": "integer",
@@ -4325,6 +4478,36 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "string",
         "description": "Service area in which plant is located; for unregulated companies, it's the electric utility with which plant is interconnected",
     },
+    "services_efv_in_system": {
+        "type": "integer",
+        "description": (
+            "Estimated number of services with Excess Flow Valve "
+            "in the system at end of reported year related to "
+            "natural gas distribution."
+        ),
+    },
+    "services_efv_installed": {
+        "type": "integer",
+        "description": (
+            "Total number of services with Excess Flow Valve installed "
+            "during reported year related to natural gas distribution."
+        ),
+    },
+    "services_shutoff_valve_in_system": {
+        "type": "integer",
+        "description": (
+            "Estimated number of services with manual service line "
+            "shut-off valves installed in the system at end of report year "
+            "related to natural gas distribution."
+        ),
+    },
+    "services_shutoff_valve_installed": {
+        "type": "integer",
+        "description": (
+            "Total number of manual service line shut-off valves installed "
+            "during reported year related to natural gas distribution."
+        ),
+    },
     "service_type": {
         "type": "string",
         "description": (
@@ -4604,6 +4787,7 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         # TODO: disambiguate the column name. State means different things in
         # different tables. E.g. state of the utility's HQ address vs. state that a
         # plant is located in vs. state in which a utility provides service.
+        # TODO: Figure out which enum to use here - include Canadian provinces?
         "description": "Two letter US state abbreviation.",
     },
     "state_id_fips": {