+++ docs/release_notes.rst
@@ -56,6 +56,13 @@ SEC 10-K
   * :ref:`out_sec10k__changelog_company_name`
   * :ref:`out_sec10k__changelog_company_name`
 
+PHMSA
+~~~~~
+* Add a transformed table containing annual operator data from PHMSA natural gas
+  distributors. This is a subset of the overall distributor data, focusing on
+  company-level attributes. Thanks to :user:`seeess1` for all of your work on this! See
+  :issue:`3770` and :pr:`4005`.
+
 Expanded Data Coverage
 ^^^^^^^^^^^^^^^^^^^^^^